Update to support new extensibility for finding Menus
https://github.com/heyday/silverstripe-menumanager/pull/54

With heyday/silverstripe-menumanager:
When creating a menu set, a validation was introduced to ensure that
no menu already existed with that name. However no allowance was
made to consider various other filters such as the currently active
subsite.

This was to combat a bug where duplicates could arise but then could
not be managed in the CMS. This change however unfortuantely broke
the integration with this module on a dev/build, where the
configured default menus were repeated for each subsite, with the
same names.

This does not affect sites that already have the menus configured
and e.g. the existing database is copied for development purposes.
But on a fresh project install without existing data this would
cause dev/build fatal/emergency level errors and prevent the
successful conclusion of the script (lots of assumed default data
could be missing). This commit removes this development annoyance.